+++ public-info/src/main/java/com/bank/publicinfo/entity/BankDetails.java
@@ -1,7 +1,6 @@
 package com.bank.publicinfo.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -20,7 +19,7 @@ public class BankDetails {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private BigInteger id;
+    private Long id;
 
     @Column(name = "bik", nullable = false)
     private BigInteger bik;
